Change Command to be an Abstract class
Since all Commands only implement the functionality defined by
the Command class, but don't provide any further functionality
themselves it makes more sense from both a modelling and code
perspective for Command to be an Abstract class. It cannot exist
without having a specific implementation, and it is a type, not a
contract.