Change Entity to an Abstract class
Since Entity is the root of all classes in the Entities package
it makes more sense for it to be a type than an interface. All
classes in Entities are subtypes of Entity, they don't provide some
implementation that other classes use - they _are_ entities and the
model should reflect this.