[UNSTABLE] Massively overhaul Entity system - changes apparent accross ALL classes
To enable a greater level of customization for modders it became
apparent that giving all entities commands, inventories, and stats
would be neccessary. This required a restructuring of almost all
classes.

As a larger benefit, the overall codebase is now much more robust
(significantly less repeated code), and the entity heirarchy is now
much simpler (this also helped get rid of some, but not all, casting).

In the state of this commit it is likely that almost all of the JSON
files are no longer valid - however it is possible to compile and run
the menus etc.